addressed review comments 3 for eventing upgradejob
