--- components/eventing-controller/upgrade-job/process/add_steps.go
@@ -1,12 +1,0 @@
-package process
-
-// AddSteps defines the execution sequence of steps for upgrade-job
-func (p *Process) AddSteps() {
-	p.Steps = []Step{
-		NewScaleDownEventingController(p),
-		NewDeletePublisherDeployment(p),
-		NewGetSubscriptions(p),
-		NewFilterSubscriptions(p),
-		NewDeleteBebSubscriptions(p),
-	}
-}
